requirements.txt
PyQtGraph
PyQt5
pyaudio
PyYAML
